+++ package.json
@@ -25,6 +25,7 @@
     "react-scripts": "3.4.0",
     "redux": "^4.0.5",
     "redux-devtools": "^3.5.0",
+    "redux-firestore": "^0.13.0",
     "redux-thunk": "^2.3.0",
     "ts-date": "^3.0.0",
     "typescript": "~3.7.2",
