+++ src/lib/firebase.ts
@@ -3,21 +3,19 @@ import 'firebase/firestore';
 import { Event } from './event';
 import { Dispatch } from 'react';
 import { EventActions } from 'redux/actions';
+import { eventCollection } from 'redux/store';
 
-const fbConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
-  authDomain: 'asia-northeast1',
-  projectId: 'todays-movie',
-};
+// const fbConfig = {
+//   apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
+//   authDomain: 'asia-northeast1',
+//   projectId: 'todays-movie',
+// };
 
-firebase.initializeApp(fbConfig);
+// firebase.initializeApp(fbConfig);
 
 export type EventDocs = firebase.firestore.QuerySnapshot<
   firebase.firestore.DocumentData
 >;
-
-const db = firebase.firestore();
-const eventCollection = db.collection('events');
 
 // export async function fetchEventDocs(
 //   successFunc: (
